file switching: sort selection dialog for recently visited files
When we switch files, users should expect more recently visited files
closer to the top of the selection dialog.  For this we make selected
files the new predecessor of the current file, effectively getting a
stack sortet by last visit.

Now, the selection dialog shows the current file as the first entry
but the initially selected file to switch to is the second line.  This
way, the user can easily stay at the previous file by entering 'q'uit
or moving the selection one line up and pressing ENTER.

Further, there was an error in file switching: when the user left the
dialog with 'q'uit (i.e. no selection) we switched to the previous
file in the ring.  This got fixed.

Signed-off-by: Dirk Gouders <dirk@gouders.net>