manual pages: improve repositioning of reloaded manual pages
When reloading manual pages we just used the prior file position:
adjusted it to the beginning of a line and used that result as the new
position from where to continue paging.

Alejandro Colomar suggested a better approach: count the words inside
the current section up to the original position and use that value for
repositioning.

To reduce hyphenation effects, I modified this to count words up to
a preceeding empty line (or the section header) and the continue to
only count empty lines until we reach the section header.  After
reload, we then forward that amount of empty lines and finally the
amount of words we counted.

This approach can be refined further: if a lot of words (to be
defined) are found, set it in relation to all the words in the section
and work with that value.  A currently unused function
lsp_man_count_words_in_section() is therefore part of this commit.

Signed-off-by: Dirk Gouders <dirk@gouders.net>