EOF detection: fix segfault with last lines of files
When the last line of a file was also displayed as the last line of a
page, further movement could cause segfaults because of incorrect
detection of EOF.

Fix this in the line getter by finally peeking one more byte forward
if the file's size is still unknown.  This will detect EOF where it
previously wasn't.

Signed-off-by: Dirk Gouders <dirk@gouders.net>