ANSI escape sequences: handle just SGR
Up to now, *lsp* was unclear in if/how to handle ANIS escape
sequences and errored out if it found some unknown sequence.

After handling some logfiles that included various escape sequences it
was clear to me that *lsp* should just try to handle SGR sequences;
the rest should be output as is.

In the long term, high level ncurses terminal handling should be
replaced by terminfo / termios and from that point even SGR will no
longer be interpreted by *lsp* but the terminal driver; *lsp* should then
still recognize SGR to enable searching in SGR intermixed text but
that's it.

Signed-off-by: Dirk Gouders <dirk@gouders.net>