Command input: just use the controlling terminal
Our prior handling of command input was ugly and didn't work, for
example, if lsp was used in a su(1)-session.  That was because the
device files ttyname(3) gives us don't belong to the new user but the
one who invokes su(1).

Clean up the code and just use the controlling terminal if we need to
open a terminal device for command input.

Signed-off-by: Dirk Gouders <dirk@gouders.net>