lsp: fix EOL matching
Search patterns containing EOL operator didn't work.

Use cflag REG_NEWLINE instead of removing newlines from the lines we
are searching in (which we didn't do everywhere).

This bug came in with the first commit and I don't remember why I
decided doing the searches that way.  Could be that reason will
require more work on regex stuff in the future...

Signed-off-by: Dirk Gouders <dirk@gouders.net>