data buffers: fix movement within file not yet fully buffered
Movement in files not yet completely read caused an error when moving
back to a position not inside the last read data buffer and then
forward to a position that causes the next read.

Fix this by ensuring we have that last read buffer at hand before
adding the next one.

Signed-off-by: Dirk Gouders <dirk@gouders.net>