lsp_line_t: add a pointer to the current position of line handling
Preparation for fixing the handling of grotty's legacy output format.

We start with trying to prepare for a cleanup of the mess called
lsp_display_page().
One reason this function is so ugly is that we maintain
function-local pointers/indexes to current positions in processed
lines which makes it hard to create functions that can perform
sub-tasks.

Add a pointer _current_ to struct lsp_line_t (and a macro that
translates it to an index) to enable us creating functions that
perform defined tasks on the current line.

Signed-off-by: Dirk Gouders <dirk@gouders.net>