open manpages: use MAN_PN to get the file names
Our prior way to find names of manual pages was to inspect the first
line of a manual page which -- in almost all cases -- should contain
the proper name twice.

But, there are cases where those names are of no use and cause real
harm to us -- docker is an example, using the name "docker(1)" for all
of its manual pages.

Luckily, man-db's man(1) provides MAN_PN containing the name of the
manual page.  But when we created childs executing man(1) we have no
access to their MAN_PN.

Try this hack to get the names of such manual pages:

* Create a script lsp_cat that outputs a heading line with the value
  of MAN_PN prior to the actual content of the manual page:

  <lsp-man-pn>lsp(1)</lsp-man-pn>

* Create child processes with PAGER=lsp_cat.

* Extract the name of the manual page from the first line of
  the incoming data.

For this, we built a more general read function lsp_file_do_read().
(Looks as if there was an error in lsp_file_read_block() anyway, by
adding the (possibly negative) result of read(2) to nread...)

Signed-off-by: Dirk Gouders <dirk@gouders.net>