open manpage: handle nonexistent manual pages
With a previous commit we error out if the user tries to open a manual
page that does not exist.  That wasn't the previous behaviour and we
want to have that back so that the user doesn't loose a session with
possibly several opened files just because of a typo or similar.

Fix this in lsp_read_manpage_name() and use a new function
lsp_file_inject_line() if we didn't find the line we expected.
This function injects the read data into cf's data buffer as if it
were read the usual way.

This also handles situations where an implementation of man(1) is used
that doesn't provide MAN_PN.  Manual pages that have problematic names
in their headers would then again cause problems, though.

Further, add a flag to lsp_detect_manpage() to only check MAN_PN for
files we build from stdin.  Otherwise, files that be build by
executing man(1) would try to use the name of the wrong manual page.

Fixes: fa5240db7874b581 (open manpages: use MAN_PN to get the file names)
Signed-off-by: Dirk Gouders <dirk@gouders.net>