regexec: don't rely on string terminators
This starts a series of patches to become more resilient to non-text
input.

We start to enforce a simple rule:

Strings are terminated by lengths and not special characters.

This one starts with regexec(3) and provides start and end positions
prior to their calls.

Another advantage is that regexec(3) no longer calls strlen(3).

Signed-off-by: Dirk Gouders <dirk@gouders.net>