lsp_error: fix a call in lsp_get_gref_at_pos()
In lsp_get_gref_at_pos() lsp_error() was called to log a problem with
a line not found at a certain position.  This call was broken, because
not the position (i.e. type off_t) but a regmatch_t was handed to
lsp_error().

So far, this call never happened and this problem has been noticed by
chance.

Signed-off-by: Dirk Gouders <dirk@gouders.net>