strings: finer distinguish from pure data
This patch is the next step in distinguishing between strings and
input data where null-characters could be part of the payload.

For this, two new functions lsp_mdup() and lsp_mdup2str() were
introduced.  The former duplicates plain data, the latter does the
same but finally builds a string appending a null-character.

One might ask: "Why not use strndup()?" and the answer is that
strndup() copies *at most* n bytes but we need copies of *exactly*
n bytes.

Further, lsp_to_lower() was modified to modify the given string
in-place and no longer return a duplicate.  This was done, because it
was more convenient to prepare a string before calling lsp_to_lower().

With all that done, we ensure that names passed to *gref*-functions
are guaranteed to be strings, gaining an uncomplicated
handling.

At other places: use the new lsp_mdup*-functions where applicable.

Signed-off-by: Dirk Gouders <dirk@gouders.net>