lsp_line_t: forget about strings
The members .raw and .normalized are no longer seen as strings.

One consequence is that some functions now need to know about their lengths.

Also, a new function lsp_normalize2str() was introduced to be able to
receive a string out of a normalization.

Unfortunately, this patch is a hughe change but I saw no comfortable
way to further brake it down.

Signed-off-by: Dirk Gouders <dirk@gouders.net>