lsp_get_line_from_here: stop replacing null-characters
This is the final (for now) step in handling any input: we do no
longer replace null-characters when preparing lines.

This procedure was broken anyway and not really protected lsp from
segfaulting.

What remains is a translation from \r to ^M and I do not yet know what
to do with those (#29)...

Further, one call to lsp_debug() needed to be fixed -- it makes no
sense to log something with %s that mblen() failed to detect as a
proper character.  Log it in hex.

Finally, complete the mess by no longer talking about strings where we
actually work on plain data.
Fix not only comments but also variable names.

Fixes #26

Signed-off-by: Dirk Gouders <dirk@gouders.net>