non-empty files: handle them correctly
An example for problems that arise when a "trivial" commit
is done without our currently evolving pre-commit testing.

Commit 2bc2241b5801 (empty files: handle them correctly) introduced a
problem with non-empty files:

The size of non-regular files is known only after we read all of it.
Since then their size is (off_t)-1 and a test for > 0 matches them
(negatively).

Test for == 0 if empty files are meant.

Fixes: 2bc2241b5801 (empty files: handle them correctly)
Signed-off-by: Dirk Gouders <dirk@gouders.net>