carriage return: revise handling
It was time to revise the code for translating carriage return --
non-text input too easily contains them.

We no longer do the translation when handing around lines but now
just before we output them to ncurses.  The former was broken and now,
searches for carriage return are possible in theory.

Being at this a new option --keep-cr was introduced to turn the
translation off.

Signed-off-by: Dirk Gouders <dirk@gouders.net>