pager: use correct environment variable
man-db's man(1) has an ordered set of environment variables that
determine the pager to use.  MANPAGER has the highest priority and we
should set it accordingly.

The fix is: when child processes are about to execute man(1), they
check if MANPAGER is set and if so prepare this variable.  PAGER is
used, otherwise.

Signed-off-by: Dirk Gouders <dirk@gouders.net>