lsp_sgr_extract_enns(): fix possible array overflow
The function lsp_sgr_extract_enns() introduced a (or one more)
security risk by not checking the given array's limit.

Fix this by handing over the size of the array and checking if i
exceeds the given size.

The error handling so far is to just return -1 and thus ignore the
complete sequence.

Signed-off-by: Dirk Gouders <dirk@gouders.net>