lsp_display_page: fix problem at end of files
The last page of files with long lines was broken: only the
beginning of a final long line was shown, because cf->page_last was
set to the end of the long line and not to the position that caused
the page being filled.

When processing lines, we need to keep in mind that they could be much
longer than the width of the window.

Fix the loop that processes a line so that it also stops when the maximum
height of the window is reached.  This results in cf->page_last being set
to the correct value.

While at it, remove a pair of superfluous parentheses.

Signed-off-by: Dirk Gouders <dirk@gouders.net>