hidden window: don't make it wider than the real one
There is no point in making the hidden window one column wider than
the visible one.  We could have characters that expand more than one
column and actually there is no problem in checking if a new character
brought us into the second line.

Fix this and make the hidden window as wide as the visible one.

Signed-off-by: Dirk Gouders <dirk@gouders.net>