add_wlines: correctly maintain current_col
After we write characters to the hidden window we need to update
current_col as well.

Fix this after getting current coordinates after a write.

Also, add an assert to the case when we moved to the 2nd line in the
hidden window, just to ensure our assumptions for that situation are
correct.

Signed-off-by: Dirk Gouders <dirk@gouders.net>