searching: implement switching of match-positioning strategy
Search matches are tried to be positioned so that context is also
visible and CTRL-l can be used to bring the current match to the top
of the window if that is wanted.

But if we are in a search where it would make sense to have _every_
match in the top line, we always need to press CTRL-l and that can
be frustrating.

Implement a switch of positioning stragegies: if CTRL-l is pressed
twice, the current strategy will be switched permanently, until CTRL-l
is pressed twice, again, etc.

Signed-off-by: Dirk Gouders <dirk@gouders.net>