grefs: Fix double-free in lsp_grefs_dtor() on macOS
On macOS hdestroy() calls free() for each key in the search table.
Because with the GNU C Library the keys are not freed (because the may
be constants) we are doing that cleanup in lsp_grefs_dtor(), causing a
double-free on macOS.

Duplicate the keys on macOS so that hdestroy() later can free the
keys without getting in our way.

Signed-off-by: Tom Vierjahn <tom.vierjahn@acm.org>
Signed-off-by: Dirk Gouders <dirk@gouders.net>