build/macOS: Make sure _DARWIN_C_SOURCE is defined
On some macs, meson seems not to be picking up the correct SDK during
`meson setup`. As a consequence, _DARWIN_C_SOURCE is not set in the
compile commands. This leads to several errors
  - unknown type name 'uint'
  - call to undeclared library function 'strdup'
  - use of undeclared identifier 'REG_STARTEND'
  - call to undeclared library function 'snprintf'
  - call to undeclared function 'strsignal'

In Apple's implementation of the standard library
  - strdup      [200112]
  - snprintf    [200112]
  - strsignal   [200809]
require at least a certain POSIX version being enabled (in brackets),
if _DARWIN_C_SOURCE is not set. If _DARWIN_C_SOURCE is set, the above
functions are declared by default.

The identifiers
  - uint
  - REG_STARTEND
are available, only if _DARWIN_C_SOURCE is defined, or if
_POSIX_C_SOURCE is undefined.

Signed-off-by: Tom Vierjahn <tom.vierjahn@acm.org>