argv: unify handling of argument vectors
We already had two functions where argument vectors were created.
Upcomming features will also require the use of argument vectors, so
it is time to unify the handling of argv.

Create functions for creation and destruction of argument vectors and
use them in the functions that built their own vectors.

Signed-off-by: Dirk Gouders <dirk@gouders.net>