lsp: introduce pinfo -- information to restart our parent
To restart the process that started us, we need to know its command
line.  Get this information and store it in the new structure
lsp_pinfo.

For now, this structure is a global one, because we only have one
parent process.  Could be that we later want this per file to be able
to restart processes that feed us data for various open files.
But currently, we are only able to open additional manual pages and we
already know how to restart man(1).

Signed-off-by: Dirk Gouders <dirk@gouders.net>