lsp: handle read strings containing backspaces
When we read search strings or names of manpages, the user might
correct typos using backspaces.  Those can land in the read strings
causing searches or loading of manual pages to incomprehensibly fail.

Until we know a better way, we fix that by processing just read
strings and emulating backspace behavior as the user would expect it.

Signed-off-by: Dirk Gouders <dirk@gouders.net>