automatic reloading: tighten the rules
The practice to automatically reload manual pages was a bit freehand:
the fact that we recognize a manual page doesn't say that it was
man(1) that fed it to us.

Tighten the rules for reloadable files by either ensuring it was
man(1) that called us with a manual page or that we called man(1)
for various reasons -- currently 'm' command, 'h' command.

Signed-off-by: Dirk Gouders <dirk@gouders.net>