lsp: become cat(1) if stdout isn't a tty
A pager is useless if it isn't acting on a terminal device.
If we are called with standard output not being a tty,
we simply replace ourselves with cat(1).

Could be, that we later want to remove options that are
meaningful only to lsp but this use-case is so unusual that
I prefer to wait for complaints.

Signed-off-by: Dirk Gouders <dirk@gouders.net>