TOC: ensure we have level 0 entries
lsp currently cannot operate on TOCs without level 0 entries.

With manual pages (the actual target for the TOC command) this should
not occur but that command can be used on any file.

Fix this by ensuring we get level 0 entries in lsp_toc_ctor():

We remember the lowest level that was generated and if this is not 0
when we are done, we shift the levels of all generated TOC entries so
that the lowest level becomes level 0.  This should be fine,
because users never have to deal with TOC levels.

Signed-off-by: Dirk Gouders <dirk@gouders.net>