lsp/TOC: handle cursor line when switching levels
Currently, switching TOC levels with the cursor line not at the top of
the page looks confusing, because the position of the cursor line
inside the page is not changed and therefore then shows differing
content.

Fix this by bringing the current cursor line to the top of the page
when switching TOC levels.

Signed-off-by: Dirk Gouders <dirk@gouders.net>