lsp_debug: write early messages to buffer (not stderr)
Writing early debug messages to stderr poisons the terminal window for
no good reason.

Change that behavior and store early messages in a buffer that then
gets written after the logfile file descriptor is initialized.

Another advantage of this is that all debug messages are kept
together and don't get spread to stderr and a log file.

Signed-off-by: Dirk Gouders <dirk@gouders.net>