lsp: refactor lsp_set_manpager() to lsp_set_pager()
We are moving forward to start other processes than man(1).
One problem would be to know what environment variable(s) a certain
program would check to use a pager.

For now, we choose the simple way and just set all such variables
that we know of.  We do not try to be smart and find out which one
would suffice.

Signed-off-by: Dirk Gouders <dirk@gouders.net>