lsp: refactor wait()ing for childs
We handle child processes in a very simple way: after we started
them, we simply read the whole file content and then waitpid() for the
child to exit.

This is not very elegant and if we begin to restart other
processes than man(1), this will be inacceptable.  Imagine, we always
fully read a `git log` for the Linux kernel sources, when the user
wants a reload to see a new commit that can only appear at the
beginning of the log.

Change that behavior by just reading an initial block on reload and
then let the positioning to the last displayed page get just the data
actually needed.

Add a child_pid to struct file_t and then wait for probably existing
childs in lsp_file_close() after we closed the file's file descriptor.

Signed-off-by: Dirk Gouders <dirk@gouders.net>