lsp/naming of functions: rename lsp_files_list() to lsp_cmd_show_files()
Another case of a bad function name: lsp_files_list() is a command
invoked by 'B' and as that its must start with  "lsp_cmd".

Further, in "files_list" verb and object are swapped.

Fix this and use the verb "show" instead of "list", because it
sounds more like presenting a dialog.

Signed-off-by: Dirk Gouders <dirk@gouders.net>