lsp/display_page: always write complete page
We are close to restarting parent processes and it is time to take
care for cases when content we reload, changed.

One such case would be when the reloaded content of a file shrank --
to zero in the extreme.  In such cases, there might be content on the
screen that disappeared and we have to show that.

Fix this in lsp_display_page(): don't check if we reached EOF to save
some cycles but always fill the whole screen.

Signed-off-by: Dirk Gouders <dirk@gouders.net>