lsp_file_reread: use lsp_file_set_pos_bol() instead of lsp_file_set_pos()
Re-reading regular files is buggy when the content changes
in a way that the old cf->page_first position doesn't stay at the
beginning of a line.

Fix that by using lsp_file_set_pos_bol() to ensure repositioning lands
at the beginning of a line.

Further, there is no need for the variable old_page_first, because
cf->page_first won't be modified by any function that we use.

Fixes: 5f7d2f9 (lsp_cmd_reload: introduce new command 'r'eload, 2025-01-03)
Signed-off-by: Dirk Gouders <dirk@gouders.net>