lsp_display_page: unhack that function - part I
lsp_display_page() isn't the most elegant part of *lsp* ;-)
Hacked to death until everything worked as expected and then only
touched for bug fixes.

Let's start to try to clean that up.

First, we introduce struct pg_ctx (not yet finished) to contain page
context that can be shared among functions.

With that data structure at hand, we are able to extract the
processing of all the lines to display in a page and delegate that
task to a new function lsp_page_process_lines() that now contains that
extracted code.

Signed-off-by: Dirk Gouders <dirk@gouders.net>