lsp/reloading: fix reloading of manual pages
The next fix of prior work:

Commit 790547d (lsp/reloading: enable 'r'eloading the output of git
commands, 2025-04-10) broke the reloading of manual pages.

In lsp_start_feeder(), an initial data block is now read regardless of
the file type.  But, for manual pages, we first need to process the
output of lsp_cat(1) with MAN_PN as its first line.

Now, we don't get the name of the manual page out of the first line of
input, because we try to read that initial line after a first block
was read;  the result is completely unpredictable but by chance was
hardly noticeable from the user's perspective, which allowed that
bug to hide for some time.

(The reason, it was noticed, are manual pages with a dot in their names,
 like sys_un.h(0P).  MAN_PN escapes that dot with a backslash which we
 do not yet correct and that backslash suddenly disappeared after
 reloading because of resizes.)

Correct this by immediately reading initial data blocks only for files
that aren't manual pages and for manual pages, call
lsp_read_manpage_name() prior to reading the initial data block.

Fixes: 790547d (lsp/reloading: enable 'r'eloading the output of git commands, 2025-04-10)
Signed-off-by: Dirk Gouders <dirk@gouders.net>