lsp_start_feeder(): fix a memory leak
lsp_start_feeder() doesn't care to probably free e_argv after it has
been used to create the feeder process.

But the main problem with the code is:

* e_argv is needed in the child only and because the child calls
  execvp(3) after e_argv is prepared, there is no problem with
  cleanup in the child code.

Move the preparation of e_argv into the child-code so that the parent
hasn't to care about if memory has to be free()'d.

Fixes: 0fd7bbe (options: implement --reload-command, 2023-03-29)
Signed-off-by: Dirk Gouders <dirk@gouders.net>