lsp_get_parent_cmd_line(): fix a memory leak
In lsp_get_parent_cmd_line(), the command used to get the parent
command line should be free()'d when it's no longer needed.

While at it, polish that short function a bit and, for example, use a
variable name that expresses we have a format string, instead of
"ps_cmd".

Fixes: 6711c00 (lsp: introduce pinfo -- information to restart our parent, 2024-07-22)
Signed-off-by: Dirk Gouders <dirk@gouders.net>