lsp/cleanup: unhack lsp_display_page() - part VI
Still the cleanup of lsp_display_page().

For dealing with SGR and grotty(1) backspace sequences, we
need to be able to look at the next three wchars in a line and
those are fetched at various locations.

Put that task into a new function lsp_page_next_three_wchars()
so that this detail is forced into better defined behavior.

Signed-off-by: Dirk Gouders <dirk@gouders.net>