lsp/cleanup: polish comments, fix superfluous braces, remove dead code
While working on the "unhack" sequence, I noticed that I should spend
the time and walk through comments and simple uglities in the
code like uncommented leftovers of previous fixes and also superfluous
braces in if-statements and loops.

Elaborate a bit more in the introductory comments in lsp.c,
fix too long lines and multi-line comments and remove superfluous
braces in if-statements and loops.  Also, remove uncommented code.

Signed-off-by: Dirk Gouders <dirk@gouders.net>