lsp: improve unicode resilience
Naivgation in unicode data is broken and triggers
an assert when moving backwards.

As a debuggable examples run the following in a 5x4 window.
Go forward one page and then back one line:

for n in $(seq 4) ; do
    printf "\xf0\x9f\x91\xa9" # woman
    printf "\xf0\x9f\x8f\xbd" # skin tone
    printf "\xe2\x80\x8d"     # zero width joiner
    printf "\xe2\x9d\xa4"     # heavy black heart
done | lsp

The reason for this is that character output and identification of
window lines is broken for situations when graphemes need more than
one column.

Fix that by handling cases when graphemes don't fit in the remainder
of a line and page and use proper wline starts and page ends.

Signed-off-by: Dirk Gouders <dirk@gouders.net>