lsp_line_add_wlines: refactor tab expansion
When we expand tabs in lsp_line_add_wlines(), we replace the tab
character in line->raw with a space so that the tab character doesn't
trigger more than one expansion.
But, because callers may further process the line, we must recover the
replaced tab character when we are done with the expansion to avoid
side-effects.

Also, when we find a tab character and its expansion with spaces will
fill the current wline, we should immediately trigger the next wline
instead of doing the expansion while already knowing the result.  This
also avoids to do the above recovery at a further place, when we
notice the line is filled.

Signed-off-by: Dirk Gouders <dirk@gouders.net>