lsp_wline_bw: refactor searching for a matching wline bol
In backward movements, lsp can fail to find the wline that is
currently the top of the page if window resizes come into play.

One problem is that we try to exactly match a given position and error
out if we fail.

Make this task more resilient and don't search for exact matches but
for wlines that start at offsets <= a given one.  This will always
succeed, because we will at least find the wline with offset 0.

Signed-off-by: Dirk Gouders <dirk@gouders.net>